fix: test_連字符檢查試驗，拆成mock試驗和整合試驗
